pyproject.toml
[project]
name = "cli-research-agent"
version = "0.1.0"
description = "CLI research agent with tool calling"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "httpx>=0.27.0",
    "lxml>=6.0.2",
    "openai>=2.30.0",
    "pathlib>=1.0.1",
    "python-dotenv>=1.0.0",
    "tavily-python>=0.3.0",
]
